backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic[email]>=2.0.0 # Ajout de [email] pour inclure email-validator
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Dépendances pour la base de données
SQLAlchemy>=2.0
psycopg2-binary
alembic>=1.9

# Dépendances pour la sécurité
passlib[bcrypt]
python-jose[cryptography]

# Outils de développement (à installer séparément ou dans le .venv globalement)
# black
# ruff
